Abstract isspace calls into isspace_helper function
isspace will consider different sets of characters as whitespace
depending on C locale. Usually this is fine, but the main problem
character is a non-breaking space (Unicode: U+00A0). This abstracted
helper method checks both isspace to be true OR for the data char to
be a NBS (typed on an OSX US keyboard with option-space.)